EXPERIMENT-2/LLM-generated-code/requirement.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.24.3
pillow==10.2.0
scikit-learn==1.3.0
matplotlib==3.7.2